added partition_columns to pd_to_gstorage

+++ gcloud_connectors/gstorage.py
@@ -36,29 +36,37 @@ class GStorageConnector:
         self.logger = logger if logger is not None else EmptyLogger()
 
     @retry((socket.timeout, requests.exceptions.ConnectionError, urllib3.exceptions.ProtocolError), tries=3, delay=2)
-    def pd_to_gstorage(self, df, bucket_name, file_name_path, tempfile_mode=True):
+    def pd_to_gstorage(self, df, bucket_name, file_name_path, tempfile_mode=True, partition_cols=None):
         """
         :param df: pandas DataFrame to be saved on GCS
         :param bucket_name: GCS bucket name
         :param file_name_path: path to save file on bucket
         :param tempfile_mode: if using a tempfile before pushing to GCS
+        :param partition_cols: columns for partitioning in order of partitions
         :return: True or error whether file is correctly saved or not
         """
-        if tempfile_mode:
-            bucket = self.service.get_bucket(bucket_name)
-            with tempfile.NamedTemporaryFile('w') as temp:
-                df.to_parquet(temp.name + '.parquet', index=False)
-                bucket.blob(file_name_path).upload_from_filename(temp.name + '.parquet',
-                                                                 content_type='application/octet-stream')
-                temp.flush()
-                os.remove(temp.name + '.parquet')
-                return True
+        if partition_cols is None:
+            if tempfile_mode:
+                bucket = self.service.get_bucket(bucket_name)
+                with tempfile.NamedTemporaryFile('w') as temp:
+                    df.to_parquet(temp.name + '.parquet', index=False)
+                    bucket.blob(file_name_path).upload_from_filename(temp.name + '.parquet',
+                                                                     content_type='application/octet-stream')
+                    temp.flush()
+                    os.remove(temp.name + '.parquet')
+                    return True
+            else:
+                # only works for the following order: gcloud CLI default, gcsfs cached token,
+                # google compute metadata service, anonymous
+                df.to_parquet(
+                    'gcs://{bucket_name}/{file_name_path}'.format(bucket_name=bucket_name, file_name_path=file_name_path),
+                    index=False)
         else:
             # only works for the following order: gcloud CLI default, gcsfs cached token,
             # google compute metadata service, anonymous
             df.to_parquet(
                 'gcs://{bucket_name}/{file_name_path}'.format(bucket_name=bucket_name, file_name_path=file_name_path),
-                index=False)
+                index=False, partition_cols=partition_cols)
 
     def recursive_delete(self, bucket_name, directory_path_to_delete):
         """
